requirements.txt
torch>=1.9.0
torchvision>=0.10.0
gradio>=3.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=8.0.0
numpy>=1.21.0
